pyproject.toml
[project]
name = "mmer-xai"
version = "0.1.0"
description = "Multimodal Emotion Recognition with Explainable AI and Natural Language Generation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.13"
authors = [{ name = "Abhi" }]

dependencies = [
    # Core DL
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    # Transformers / NLP
    "transformers>=4.36.0",
    "tokenizers>=0.15.0",
    # Vision
    "timm>=0.9.12",
    "Pillow>=10.0.0",
    # XAI
    "shap>=0.43.0",
    "captum>=0.7.0",
    "grad-cam>=1.5.0",
    # Data & Science
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Config & Logging
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
    "tqdm>=4.65.0",
    # Video Processing (MELD)
    "opencv-python>=4.8.0",
    # Utilities
    "requests>=2.31.0",
    "gdown>=4.7.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
wandb = ["wandb>=0.15.0"]
llm = ["ollama>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data", "models", "explainers", "training", "utils", "scripts"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ipykernel>=6.0",
]
